Split build logic out
