+++ packages/blade/public/universal/utils.ts
@@ -1,0 +1,12 @@
+export interface SourceFile {
+  path: string;
+  content: string;
+}
+
+export interface BuildOptions {
+  sourceFiles: Array<SourceFile>;
+}
+
+export const build = (options: BuildOptions) => {
+  console.log('test');
+};
